♻️ [devext] move components modules around
