+++ developer-extension/src/panel/components/panel.tsx
@@ -1,13 +1,13 @@
 import React, { useState } from 'react'
 import { Tabs, Text } from '@mantine/core'
 
-import type { Settings } from './components/settingsTab'
-import { SettingsTab } from './components/settingsTab'
-import { InfosTab } from './components/infosTab'
-import { useEvents } from './hooks/useEvents'
-import { EventTab } from './components/eventsTab'
-import { useStore } from './hooks/useStore'
-import { useAutoFlushEvents } from './hooks/useAutoFlushEvents'
+import { useEvents } from '../hooks/useEvents'
+import { useStore } from '../hooks/useStore'
+import { useAutoFlushEvents } from '../hooks/useAutoFlushEvents'
+import type { Settings } from './tabs/settingsTab'
+import { SettingsTab } from './tabs/settingsTab'
+import { InfosTab } from './tabs/infosTab'
+import { EventTab } from './tabs/eventsTab'
 
 const enum PanelTabs {
   Events = 'events',
